test(functional): move model tests under test groups
