+++ tests/functional_tests/launch_scripts/h100/active/L0_Launch_models_qwen3_omni.sh
@@ -1,3 +1,4 @@
+# CI_TIMEOUT=50
 #!/bin/bash
 # Copyright (c) 2026, NVIDIA CORPORATION.  All rights reserved.
 #
@@ -15,10 +16,10 @@
 
 set -xeuo pipefail
 
-REPO_ROOT=$(cd "$(dirname "$0")/../../.." && pwd)
+REPO_ROOT=$(cd "$(dirname "$0")/../../../../.." && pwd)
 export CUDA_VISIBLE_DEVICES="${CUDA_VISIBLE_DEVICES:-0,1}"
 
 uv run coverage run --data-file="${REPO_ROOT}/.coverage" --source="${REPO_ROOT}" --parallel-mode -m pytest \
   -o log_cli=true -o log_cli_level=INFO -v -s -x -m "not pleasefixme" --tb=short -rA \
-  tests/functional_tests/models/qwen_omni/test_qwen3_omni_conversion.py
+  tests/functional_tests/test_groups/models/qwen3_omni
 coverage combine -q
