+++ tests/functional_tests/test_groups/models/qwen3_omni/__init__.py
